+++ src/main/java/com/github/danielwegener/logback/kafka/KafkaAppenderConfig.java
@@ -1,18 +1,17 @@
 package com.github.danielwegener.logback.kafka;
 
 import ch.qos.logback.core.UnsynchronizedAppenderBase;
+import ch.qos.logback.core.encoder.Encoder;
 import ch.qos.logback.core.spi.AppenderAttachable;
 import com.github.danielwegener.logback.kafka.delivery.AsynchronousDeliveryStrategy;
 import com.github.danielwegener.logback.kafka.delivery.DeliveryStrategy;
-import com.github.danielwegener.logback.kafka.encoding.KafkaMessageEncoder;
 import com.github.danielwegener.logback.kafka.keying.KeyingStrategy;
-import com.github.danielwegener.logback.kafka.keying.RoundRobinKeyingStrategy;
-import static org.apache.kafka.clients.producer.ProducerConfig.*;
+import com.github.danielwegener.logback.kafka.keying.NoKeyKeyingStrategy;
 
 import java.util.HashMap;
-import java.util.HashSet;
 import java.util.Map;
-import java.util.Set;
+
+import static org.apache.kafka.clients.producer.ProducerConfig.*;
 
 /**
  * @since 0.0.1
@@ -21,53 +20,18 @@ public abstract class KafkaAppenderConfig<E> extends UnsynchronizedAppenderBase<
 
     protected String topic = null;
 
-    protected KafkaMessageEncoder<E> encoder = null;
+    protected Encoder<E> encoder = null;
     protected KeyingStrategy<? super E> keyingStrategy = null;
     protected DeliveryStrategy deliveryStrategy;
 
-    public static final Set<String> KNOWN_PRODUCER_CONFIG_KEYS = new HashSet<String>();
-    public static final Map<String,String> DEPRECATED_PRODUCER_CONFIG_KEYS = new HashMap<String, String>();
-    static {
-        KNOWN_PRODUCER_CONFIG_KEYS.add(BOOTSTRAP_SERVERS_CONFIG);
-        KNOWN_PRODUCER_CONFIG_KEYS.add(METADATA_FETCH_TIMEOUT_CONFIG);
-        DEPRECATED_PRODUCER_CONFIG_KEYS.put(METADATA_FETCH_TIMEOUT_CONFIG, MAX_BLOCK_MS_CONFIG);
-        KNOWN_PRODUCER_CONFIG_KEYS.add(METADATA_MAX_AGE_CONFIG);
-        KNOWN_PRODUCER_CONFIG_KEYS.add(BATCH_SIZE_CONFIG);
-        KNOWN_PRODUCER_CONFIG_KEYS.add(BUFFER_MEMORY_CONFIG);
-        KNOWN_PRODUCER_CONFIG_KEYS.add(ACKS_CONFIG);
-        KNOWN_PRODUCER_CONFIG_KEYS.add(TIMEOUT_CONFIG);
-        DEPRECATED_PRODUCER_CONFIG_KEYS.put(METADATA_FETCH_TIMEOUT_CONFIG, REQUEST_TIMEOUT_MS_CONFIG);
-        KNOWN_PRODUCER_CONFIG_KEYS.add(LINGER_MS_CONFIG);
-        KNOWN_PRODUCER_CONFIG_KEYS.add(CLIENT_ID_CONFIG);
-        KNOWN_PRODUCER_CONFIG_KEYS.add(SEND_BUFFER_CONFIG);
-        KNOWN_PRODUCER_CONFIG_KEYS.add(RECEIVE_BUFFER_CONFIG);
-        KNOWN_PRODUCER_CONFIG_KEYS.add(MAX_REQUEST_SIZE_CONFIG);
-        KNOWN_PRODUCER_CONFIG_KEYS.add(RECONNECT_BACKOFF_MS_CONFIG);
-        KNOWN_PRODUCER_CONFIG_KEYS.add(BLOCK_ON_BUFFER_FULL_CONFIG);
-        DEPRECATED_PRODUCER_CONFIG_KEYS.put(METADATA_FETCH_TIMEOUT_CONFIG, null);
-        KNOWN_PRODUCER_CONFIG_KEYS.add(RETRIES_CONFIG);
-        KNOWN_PRODUCER_CONFIG_KEYS.add(RETRY_BACKOFF_MS_CONFIG);
-        KNOWN_PRODUCER_CONFIG_KEYS.add(COMPRESSION_TYPE_CONFIG);
-        KNOWN_PRODUCER_CONFIG_KEYS.add(METRICS_SAMPLE_WINDOW_MS_CONFIG);
-        KNOWN_PRODUCER_CONFIG_KEYS.add(METRICS_NUM_SAMPLES_CONFIG);
-        KNOWN_PRODUCER_CONFIG_KEYS.add(METRIC_REPORTER_CLASSES_CONFIG);
-        KNOWN_PRODUCER_CONFIG_KEYS.add(MAX_IN_FLIGHT_REQUESTS_PER_CONNECTION);
-        KNOWN_PRODUCER_CONFIG_KEYS.add(KEY_SERIALIZER_CLASS_CONFIG);
-        KNOWN_PRODUCER_CONFIG_KEYS.add(VALUE_SERIALIZER_CLASS_CONFIG);
-        KNOWN_PRODUCER_CONFIG_KEYS.add(CONNECTIONS_MAX_IDLE_MS_CONFIG);
-        KNOWN_PRODUCER_CONFIG_KEYS.add(PARTITIONER_CLASS_CONFIG);
-        KNOWN_PRODUCER_CONFIG_KEYS.add(MAX_BLOCK_MS_CONFIG);
-        KNOWN_PRODUCER_CONFIG_KEYS.add(REQUEST_TIMEOUT_MS_CONFIG);
-    }
+    protected Integer partition = null;
+
+    protected boolean appendTimestamp = true;
 
     protected Map<String,Object> producerConfig = new HashMap<String, Object>();
 
-
-
     protected boolean checkPrerequisites() {
         boolean errorFree = true;
-
-
 
         if (producerConfig.get(BOOTSTRAP_SERVERS_CONFIG) == null) {
             addError("No \"" + BOOTSTRAP_SERVERS_CONFIG + "\" set for the appender named [\""
@@ -86,20 +50,20 @@ public abstract class KafkaAppenderConfig<E> extends UnsynchronizedAppenderBase<
         }
 
         if (keyingStrategy == null) {
-            addInfo("No partitionStrategy set for the appender named [\"" + name + "\"]. Using default RoundRobin strategy.");
-            keyingStrategy = new RoundRobinKeyingStrategy();
+            addInfo("No explicit keyingStrategy set for the appender named [\"" + name + "\"]. Using default NoKeyKeyingStrategy.");
+            keyingStrategy = new NoKeyKeyingStrategy();
         }
 
         if (deliveryStrategy == null) {
-            addInfo("No sendStrategy set for the appender named [\""+name+"\"]. Using default asynchronous strategy.");
+            addInfo("No explicit deliveryStrategy set for the appender named [\""+name+"\"]. Using default asynchronous strategy.");
             deliveryStrategy = new AsynchronousDeliveryStrategy();
         }
 
         return errorFree;
     }
 
-    public void setEncoder(KafkaMessageEncoder<E> layout) {
-        this.encoder = layout;
+    public void setEncoder(Encoder<E> encoder) {
+        this.encoder = encoder;
     }
 
     public void setTopic(String topic) {
@@ -116,20 +80,7 @@ public abstract class KafkaAppenderConfig<E> extends UnsynchronizedAppenderBase<
             addProducerConfigValue(split[0], split[1]);
     }
 
-
     public void addProducerConfigValue(String key, Object value) {
-        if (!KNOWN_PRODUCER_CONFIG_KEYS.contains(key))
-            addWarn("The key \""+key+"\" is not a known kafka producer config key.");
-
-        if (DEPRECATED_PRODUCER_CONFIG_KEYS.containsKey(key)) {
-            final StringBuilder deprecationMessage =
-                    new StringBuilder("The key \""+key+"\" is deprecated in kafka and may be removed in a future version.");
-            if (DEPRECATED_PRODUCER_CONFIG_KEYS.get(key) != null) {
-                deprecationMessage.append(" Consider using key \"").append(DEPRECATED_PRODUCER_CONFIG_KEYS.get(key)).append("\" instead.");
-            }
-            addWarn(deprecationMessage.toString());
-        }
-
         this.producerConfig.put(key,value);
     }
 
@@ -141,6 +92,16 @@ public abstract class KafkaAppenderConfig<E> extends UnsynchronizedAppenderBase<
         this.deliveryStrategy = deliveryStrategy;
     }
 
+    public void setPartition(Integer partition) {
+        this.partition = partition;
+    }
 
+    public boolean isAppendTimestamp() {
+        return appendTimestamp;
+    }
+
+    public void setAppendTimestamp(boolean appendTimestamp) {
+        this.appendTimestamp = appendTimestamp;
+    }
 
 }