+++ src/test/java/com/github/danielwegener/logback/kafka/KafkaAppenderTest.java
@@ -1,36 +1,43 @@
 package com.github.danielwegener.logback.kafka;
 
-import ch.qos.logback.classic.Level;
-import ch.qos.logback.classic.LoggerContext;
-import ch.qos.logback.classic.spi.ILoggingEvent;
-import ch.qos.logback.classic.spi.LoggingEvent;
-import ch.qos.logback.core.BasicStatusManager;
-import ch.qos.logback.core.status.ErrorStatus;
-import com.github.danielwegener.logback.kafka.delivery.DeliveryStrategy;
-import com.github.danielwegener.logback.kafka.delivery.FailedDeliveryCallback;
-import com.github.danielwegener.logback.kafka.encoding.KafkaMessageEncoder;
-import com.github.danielwegener.logback.kafka.keying.KeyingStrategy;
-import org.apache.kafka.clients.producer.KafkaProducer;
-import org.apache.kafka.clients.producer.ProducerRecord;
-import org.junit.After;
-import org.junit.Before;
-import org.junit.Test;
-
 import static org.hamcrest.Matchers.empty;
+import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.hasItem;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertThat;
 import static org.junit.Assert.assertTrue;
 import static org.mockito.Matchers.any;
 import static org.mockito.Matchers.eq;
-import static org.mockito.Mockito.*;
+import static org.mockito.Matchers.same;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.verifyZeroInteractions;
+import static org.mockito.Mockito.when;
+import static org.mockito.Mockito.never;
+
+import ch.qos.logback.classic.Level;
+import ch.qos.logback.classic.LoggerContext;
+import ch.qos.logback.classic.spi.ILoggingEvent;
+import ch.qos.logback.classic.spi.LoggingEvent;
+import ch.qos.logback.core.BasicStatusManager;
+import ch.qos.logback.core.encoder.Encoder;
+import ch.qos.logback.core.status.ErrorStatus;
+import com.github.danielwegener.logback.kafka.delivery.DeliveryStrategy;
+import com.github.danielwegener.logback.kafka.delivery.FailedDeliveryCallback;
+import com.github.danielwegener.logback.kafka.keying.KeyingStrategy;
+import org.apache.kafka.clients.producer.KafkaProducer;
+import org.apache.kafka.clients.producer.ProducerRecord;
+import org.junit.After;
+import org.junit.Before;
+import org.junit.Test;
+import org.mockito.ArgumentCaptor;
 
 public class KafkaAppenderTest {
 
-    private final KafkaAppender<ILoggingEvent> unit = new KafkaAppender<ILoggingEvent>();
+    private final KafkaAppender<ILoggingEvent> unit = new KafkaAppender<>();
     private final LoggerContext ctx = new LoggerContext();
     @SuppressWarnings("unchecked")
-    private final KafkaMessageEncoder<ILoggingEvent> encoder =  mock(KafkaMessageEncoder.class);
+    private final Encoder<ILoggingEvent> encoder =  mock(Encoder.class);
     private final KeyingStrategy<ILoggingEvent> keyingStrategy =  mock(KeyingStrategy.class);
     @SuppressWarnings("unchecked")
     private final DeliveryStrategy deliveryStrategy =  mock(DeliveryStrategy.class);
@@ -75,7 +82,7 @@ public class KafkaAppenderTest {
     }
 
     @Test
-    public void testDontStartWithoutBoostrapServers() {
+    public void testDontStartWithoutBootstrapServers() {
         unit.getProducerConfig().clear();
         unit.start();
         assertFalse("isStarted", unit.isStarted());
@@ -93,21 +100,38 @@ public class KafkaAppenderTest {
     }
 
     @Test
-    public void testAppend() {
-        when(encoder.doEncode(any(ILoggingEvent.class))).thenReturn(new byte[]{0x00, 0x00});
+    public void testAppendUsesKeying() {
+        when(encoder.encode(any(ILoggingEvent.class))).thenReturn(new byte[]{0x00, 0x00});
         unit.start();
         final LoggingEvent evt = new LoggingEvent("fqcn",ctx.getLogger("logger"), Level.ALL, "message", null, new Object[0]);
         unit.append(evt);
         verify(deliveryStrategy).send(any(KafkaProducer.class), any(ProducerRecord.class), eq(evt), any(FailedDeliveryCallback.class));
+        verify(keyingStrategy).createKey(same(evt));
+        verify(deliveryStrategy).send(any(KafkaProducer.class), any(ProducerRecord.class), eq(evt), any(FailedDeliveryCallback.class));
+    }
+
+    @Test
+    public void testAppendUsesPreConfiguredPartition() {
+        when(encoder.encode(any(ILoggingEvent.class))).thenReturn(new byte[]{0x00, 0x00});
+        ArgumentCaptor<ProducerRecord> producerRecordCaptor = ArgumentCaptor.forClass(ProducerRecord.class);
+        unit.setPartition(1);
+        unit.start();
+        final LoggingEvent evt = new LoggingEvent("fqcn", ctx.getLogger("logger"), Level.ALL, "message", null, new Object[0]);
+        unit.append(evt);
+        verify(deliveryStrategy).send(any(KafkaProducer.class), producerRecordCaptor.capture(), eq(evt), any(FailedDeliveryCallback.class));
+        final ProducerRecord value = producerRecordCaptor.getValue();
+        assertThat(value.partition(), equalTo(1));
     }
 
     @Test
     public void testDeferredAppend() {
-        when(encoder.doEncode(any(ILoggingEvent.class))).thenReturn(new byte[]{0x00, 0x00});
+        when(encoder.encode(any(ILoggingEvent.class))).thenReturn(new byte[]{0x00, 0x00});
         unit.start();
         final LoggingEvent deferredEvent = new LoggingEvent("fqcn",ctx.getLogger("org.apache.kafka.clients.logger"), Level.ALL, "deferred message", null, new Object[0]);
         unit.doAppend(deferredEvent);
-        verify(deliveryStrategy, times(0)).send(any(KafkaProducer.class), any(ProducerRecord.class), eq(deferredEvent), any(FailedDeliveryCallback.class));
+
+        verify(deliveryStrategy, never()).send(any(KafkaProducer.class), any(ProducerRecord.class), eq(deferredEvent), any(FailedDeliveryCallback.class));
+
         final LoggingEvent evt = new LoggingEvent("fqcn",ctx.getLogger("logger"), Level.ALL, "message", null, new Object[0]);
         unit.doAppend(evt);
         verify(deliveryStrategy).send(any(KafkaProducer.class), any(ProducerRecord.class), eq(deferredEvent), any(FailedDeliveryCallback.class));