+++ src/test/java/com/github/danielwegener/logback/kafka/util/EmbeddedKafkaCluster.java
@@ -1,41 +1,37 @@
 package com.github.danielwegener.logback.kafka.util;
 
+import java.util.HashMap;
+import java.util.Map;
+import kafka.metrics.KafkaMetricsReporter;
 import kafka.server.KafkaConfig;
 import kafka.server.KafkaServer;
+import org.apache.kafka.common.utils.Time;
 import scala.Some;
 
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
-import java.util.Properties;
+import scala.collection.immutable.Vector$;
 
 public class EmbeddedKafkaCluster {
     private final List<Integer> ports;
     private final String zkConnection;
-    private final Properties baseProperties;
+    private final Map<String, String> baseProperties;
 
     private final String brokerList;
 
     private final List<KafkaServer> brokers;
     private final List<File> logDirs;
 
-    public EmbeddedKafkaCluster(String zkConnection) {
-        this(zkConnection, new Properties());
-    }
 
-    public EmbeddedKafkaCluster(String zkConnection, Properties baseProperties) {
-        this(zkConnection, baseProperties, Collections.singletonList(-1));
-    }
-
-    public EmbeddedKafkaCluster(String zkConnection, Properties baseProperties, List<Integer> ports) {
+    public EmbeddedKafkaCluster(String zkConnection, Map<String, String> baseProperties, List<Integer> ports) {
         this.zkConnection = zkConnection;
         this.ports = resolvePorts(ports);
         this.baseProperties = baseProperties;
 
-        this.brokers = new ArrayList<KafkaServer>();
-        this.logDirs = new ArrayList<File>();
+        this.brokers = new ArrayList<>();
+        this.logDirs = new ArrayList<>();
 
         this.brokerList = constructBrokerList(this.ports);
     }
@@ -71,14 +67,15 @@ public class EmbeddedKafkaCluster {
             Integer port = ports.get(i);
             File logDir = TestUtils.constructTempDir("kafka-local");
 
-            Properties properties = new Properties();
+            Map<String, String> properties = new HashMap<>();
             properties.putAll(baseProperties);
-            properties.setProperty("zookeeper.connect", zkConnection);
-            properties.setProperty("broker.id", String.valueOf(i + 1));
-            properties.setProperty("host.name", "localhost");
-            properties.setProperty("port", Integer.toString(port));
-            properties.setProperty("log.dir", logDir.getAbsolutePath());
-            properties.setProperty("log.flush.interval.messages", String.valueOf(1));
+            properties.put("zookeeper.connect", zkConnection);
+            properties.put("broker.id", String.valueOf(i + 1));
+            properties.put("host.name", "localhost");
+            properties.put("port", Integer.toString(port));
+            properties.put("log.dir", logDir.getAbsolutePath());
+            properties.put("log.flush.interval.messages", String.valueOf(1));
+            properties.put("advertised.host.name", "localhost");
 
             KafkaServer broker = startBroker(properties);
 
@@ -88,18 +85,11 @@ public class EmbeddedKafkaCluster {
     }
 
 
-    private KafkaServer startBroker(Properties props) {
-        KafkaServer server = new KafkaServer(new KafkaConfig(props), new SystemTime(), Some.apply("embedded-kafka-cluster"));
+    private KafkaServer startBroker(Map<String, String> props) {
+        KafkaServer server = new KafkaServer(new KafkaConfig(props), Time.SYSTEM,
+                Some.apply("embedded-kafka-cluster"), Vector$.MODULE$.<KafkaMetricsReporter>empty());
         server.startup();
         return server;
-    }
-
-    public Properties getProps() {
-        Properties props = new Properties();
-        props.putAll(baseProperties);
-        props.put("metadata.broker.list", brokerList);
-        props.put("zookeeper.connect", zkConnection);
-        return props;
     }
 
     public String getBrokerList() {
@@ -118,6 +108,7 @@ public class EmbeddedKafkaCluster {
         for (KafkaServer broker : brokers) {
             try {
                 broker.shutdown();
+                broker.awaitShutdown();
             } catch (Exception e) {
                 e.printStackTrace();
             }