+++ CHANGELOG.md
@@ -1,10 +1,18 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-## [Unreleased]
+## [Unreleased 0.2.0]
 ### Changed
 - Fixed typos in logback warning messages emitted by `com.github.danielwegener.logback.kafka.KafkaAppenderConfig` (#28)
-- Switched default delivery strategy to `com.github.danielwegener.logback.kafka.delivery.AsynchronousDeliveryStrategy` as it is the [more sensible default](https://github.com/danielwegener/logback-kafka-appender/pull/32). 
+- Switched default delivery strategy to `com.github.danielwegener.logback.kafka.delivery.AsynchronousDeliveryStrategy` as it is the [more sensible default](https://github.com/danielwegener/logback-kafka-appender/pull/32).
+- Make artifact osgi compatible (#40)
+- Upgrade to logback 1.2.3. Removed `KafkaMessageEncoder` infrastructure in favor of `ch.qos.logback.core.encoder.Encoder` (#51).
+- Upgrade to kafka 1.0.0 (which supports protocol negotiation for older broker versions). Add use the actual log event timestamp to the kafka record.
+- Target partition can now be configured with a fixed number
+- Deprecating `BlockingDeliveryStrategy` in favor of `AsynchronousDeliveryStrategy`
+- `RoundRobinKeyingStrategy` has been renamed to `NoKeyKeyingStrategy.`
+- Semantic checking of `kafkaProducerValues` has been removed (in order to support multiple versions of kafka-clients).
+- Minimum Java version is 1.7
 
 ## [0.1.0] - 2016-02-07
 ### Changed